+++ lib/samanage/api/requester.rb
@@ -1,7 +1,7 @@
 module Samanage
 	class Api
 		# Get requester from value (email)
-		def get_requester_id(value: nil)
+		def get_requester_id(value: )
 			api_call = self.execute(path: "requesters.json?name=#{value}")
 			api_call[:data].size == 1 ? api_call[:data][0] : nil
 		end